Moved CourseReviewComponent
