+++ app/components/candidate_interface/course_review_component.html.erb
