+++ app/views/candidate_interface/application_form/_review_unsubmitted.html.erb
@@ -1,6 +1,6 @@
 <section class="govuk-!-margin-bottom-8">
   <h2 class="govuk-heading-m govuk-!-font-size-27"><%= t('section_groups.course') %></h2>
-  <%= render(CandidateInterface::ContinuousApplications::CourseReviewComponent.new(application_choice: @application_choice)) %>
+  <%= render(CandidateInterface::CourseReviewComponent.new(application_choice: @application_choice)) %>
 </section>
 
 <section class="govuk-!-margin-bottom-8">
